+++ CBIRProjekt/CBIR_Project/src/iconn/htw/main/IRandomFunction.java
@@ -2,6 +2,6 @@ package iconn.htw.main;
 
 public interface IRandomFunction {
 
-	float nextNumber();
+	double nextNumber();
 
 }
